botscode/requirements.txt
Telethon==1.19.0
Pillow==8.1.0